requirements.txt
FastAPI
SQLAlchemy
mysql-connector-python
python-dotenv